refactor: move jackson config to shared and align clue classes under game
