+++ src/main/java/com/codenames/codenames/backend/game/domain/GameManagerFactory.java
@@ -1,6 +1,6 @@
 package com.codenames.codenames.backend.game.domain;
 
-import com.codenames.codenames.backend.clue.ClueValidationService;
+import com.codenames.codenames.backend.game.application.ClueValidationService;
 import com.codenames.codenames.backend.game.application.CardGenerator;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.lobby.domain.Team;
