+++ src/main/java/com/codenames/codenames/backend/shared/config/JacksonConfig.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.config;
+package com.codenames.codenames.backend.shared.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
